lineapp/requirements.txt
flask
line-bot-sdk
google-cloud-datastore
gradio_client
